move displayable input logic to dedicated component
